For EXP-2012: Move tests for null_client into their own file
