+++ components/nimbus/src/client/mod.rs
@@ -4,7 +4,7 @@
 
 mod fs_client;
 pub(crate) mod http_client;
-mod null_client;
+pub(crate) mod null_client;
 use crate::error::{NimbusError, Result};
 use crate::Experiment;
 use crate::RemoteSettingsConfig;
