+++ components/nimbus/src/client/null_client.rs
@@ -24,23 +24,3 @@ impl SettingsClient for NullClient {
         Ok(Default::default())
     }
 }
-
-#[cfg(feature = "rkv-safe-mode")]
-#[test]
-fn test_null_client() -> Result<()> {
-    use crate::NimbusClient;
-    use tempdir::TempDir;
-
-    let _ = env_logger::try_init();
-
-    let tmp_dir = TempDir::new("test_null_client-test_null")?;
-
-    let aru = Default::default();
-    let client = NimbusClient::new(Default::default(), tmp_dir.path(), None, aru)?;
-    client.fetch_experiments()?;
-    client.apply_pending_experiments()?;
-
-    let experiments = client.get_all_experiments()?;
-    assert_eq!(experiments.len(), 0);
-    Ok(())
-}
